requirements.txt
streamlit
pandas
numpy
matplotlib
Pillow
plotly
scipy
